extended BlockStyler and added BlockMoved action
now BlockStyler operations are split so i can draw a block without needing a model, now i can remove the model reference from the graphicsObject.

now the block moving operation is done on the graphicsObject then at the end of the drag it is applied to the model, this is more transactional, the graphicsObject shouldn't modify the model directly.